requirements.txt
numpy
opencv-python
mediapipe
pyserial
